+++ cli/cli/App.cs
@@ -548,9 +548,6 @@ public class App
 		// FEDERATION COMMANDS
 		Commands.AddRootCommand<FederationCommand>();
 		Commands.AddSubCommand<ListFederationsCommand, ListServicesCommandArgs, FederationCommand>();
-		Commands.AddSubCommand<AddFederationCommand, AddFederationCommandArgs, FederationCommand>();
-		Commands.AddSubCommand<SetAllFederationsCommand, SetAllFederationsCommandArgs, FederationCommand>();
-		Commands.AddSubCommand<RemoveFederationCommand, RemoveFederationCommandArgs, FederationCommand>();
 		Commands.AddSubCommand<DisableFederationCommand, DisableFederationCommandArgs, FederationCommand>();
 		Commands.AddSubCommand<EnableFederationCommand, DisableFederationCommandArgs, FederationCommand>();
 		Commands.AddSubCommand<GetLocalRoutingKeyCommand, GetLocalRoutingKeyCommandArgs, FederationCommand>();