+++ client/Packages/com.beamable/Editor/Server/UI/Configs/BeamFederationSetting.cs
@@ -1,6 +1,8 @@
 using Beamable.Editor.BeamCli.Commands;
 using Beamable.Server.Editor.Usam;
 using System;
+using System.Collections.Generic;
+using System.Linq;
 using UnityEditor;
 using UnityEngine;
 
@@ -10,12 +12,12 @@ namespace Beamable.Editor.Microservice.UI2.Configs
 	public class BeamFederationSetting
 	{
 		// UI-state can live here.
-		public BeamFederationEntry entry = new BeamFederationEntry();
+		public List<BeamFederationEntry> entries = new ();
 
 		public bool IsValid()
 		{
-			if (entry == null) return false;
-			if (string.IsNullOrEmpty(entry.federationId)) return false;
+			if (entries == null) return false;
+			if (entries.Any(item => string.IsNullOrEmpty(item.federationId))) return false;
 
 			return true;
 		}
@@ -25,57 +27,147 @@ namespace Beamable.Editor.Microservice.UI2.Configs
 	[CustomPropertyDrawer(typeof(BeamFederationSetting))]
 	public class FederationEntryPropertyDraw : PropertyDrawer
 	{
-		private const int warningHeight = 20;
-		private float _dynamicHeight =  EditorGUIUtility.singleLineHeight;
+		private const string NO_FEDEDERATIONS_FOUND_INFO = "No Federation Implementation found. Make sure that your Microservice is implementing any federation interface with your Federation ID. You can create a new federation ID one by clicking on <b>Create->Federation Id</b>.";
+		private Dictionary<string, bool> _foldoutStates = new();
 
 		public override float GetPropertyHeight(SerializedProperty property, GUIContent label)
 		{
-			return _dynamicHeight;
+			float height = EditorGUIUtility.singleLineHeight;
+			
+			if (!_foldoutStates.TryGetValue(property.propertyPath, out bool state) || !state)
+				return height;
+
+			height += EditorGUIUtility.singleLineHeight;
+			
+			var groupedFederations = GroupFederations(property);
+
+			if (groupedFederations.Count == 0)
+			{
+				var noFedInfoContent = new GUIContent(NO_FEDEDERATIONS_FOUND_INFO);
+				GUIStyle guiStyle = new GUIStyle(EditorStyles.label)
+				{
+					wordWrap = true,
+					richText = true
+				};
+				float indentPaddingSpace = 15f;
+				height += guiStyle.CalcHeight(noFedInfoContent, EditorGUIUtility.currentViewWidth - EditorGUI.indentLevel * indentPaddingSpace);
+				return height;
+			}
+
+			height += groupedFederations.Select(kvp =>
+			{
+				float groupHeight = EditorGUIUtility.singleLineHeight; // Federation ID foldout
+				if (_foldoutStates.TryGetValue(kvp.Key, out bool foldoutState) && foldoutState)
+				{
+					groupHeight += EditorGUIUtility.singleLineHeight; // Interfaces that uses it title
+					groupHeight += groupedFederations.Values.Count * EditorGUIUtility.singleLineHeight; // For each interface
+				}
+
+				return groupHeight;
+			}).Sum();
+
+			return height;
 		}
 
 		public override void OnGUI(Rect position, SerializedProperty property, GUIContent label)
 		{
-			var usam = BeamEditorContext.Default.ServiceScope.GetService<UsamService>();
-
-			if (usam?.latestManifest == null)
+			EditorGUI.BeginProperty(position, label, property);
+			
+			var boldLabel = new GUIStyle(EditorStyles.foldout)
 			{
+				fontStyle = FontStyle.Bold
+			};
+
+			position.height = EditorGUIUtility.singleLineHeight;
+			
+			_foldoutStates.TryAdd(property.propertyPath, true);
+			_foldoutStates[property.propertyPath] = EditorGUI.Foldout(position, _foldoutStates[property.propertyPath], label, true, boldLabel);
+
+			if (!_foldoutStates[property.propertyPath])
+			{
+				EditorGUI.EndProperty();
+				return;
+			}
+			
+			position.y += EditorGUIUtility.singleLineHeight;
+			EditorGUI.indentLevel++;
+
+			var groupedFederations = GroupFederations(property);
+
+			if (groupedFederations.Count == 0)
+			{
+				var noFederationContent = new GUIContent(NO_FEDEDERATIONS_FOUND_INFO);
+				GUIStyle guiStyle = new GUIStyle(EditorStyles.label)
+				{
+					wordWrap = true,
+					richText = true,
+				};
+				position.height = guiStyle.CalcHeight(noFederationContent, EditorGUI.IndentedRect(position).width);
+				EditorGUI.LabelField(position, noFederationContent, guiStyle);
+				EditorGUI.indentLevel--;
+				EditorGUI.EndProperty();
 				return;
 			}
 
-			var options = BeamableMicroservicesSettings.availableFederationIds;
-			_dynamicHeight = EditorGUIUtility.singleLineHeight;
-			if (options.Count == 0)
+			foreach ((string federationId, List<string> interfaces) in groupedFederations)
 			{
-				_dynamicHeight += warningHeight;
-				EditorGUI.HelpBox(position, "A Federation Id is required to add federations to existing service. You can create a new one by clicking on Create->Federation Id.",MessageType.Warning);
+				string foldoutKey = federationId + property.propertyPath;
+				
+				_foldoutStates.TryAdd(foldoutKey, true);
+				position.height = EditorGUIUtility.singleLineHeight;
+				_foldoutStates[foldoutKey] = EditorGUI.Foldout(
+					position,
+					_foldoutStates[foldoutKey],
+					federationId,
+					true);
+
+				position.y += EditorGUIUtility.singleLineHeight;
+
+				if (!_foldoutStates[foldoutKey])
+				{
+					continue;
+				}
+				
+				EditorGUI.indentLevel++;
+				EditorGUI.LabelField(position,"Interfaces that uses it:", EditorStyles.boldLabel);
+				position.y += EditorGUIUtility.singleLineHeight;
+				foreach (string interfaceName in interfaces)
+				{
+					position.height = EditorGUIUtility.singleLineHeight;
+					EditorGUI.LabelField(position, $" - {interfaceName}");
+					position.y += EditorGUIUtility.singleLineHeight;
+				}
+
+				EditorGUI.indentLevel--;
 			}
-			else
-			{
-				var federationIdProp =
-					property.FindPropertyRelative(
-						$"{nameof(BeamFederationSetting.entry)}.{nameof(BeamFederationEntry.federationId)}");
-				var federationTypeProp =
-					property.FindPropertyRelative(
-						$"{nameof(BeamFederationSetting.entry)}.{nameof(BeamFederationEntry.interfaceName)}");
+			EditorGUI.indentLevel--;
+			EditorGUI.EndProperty();
+		}
 
-				var width = EditorGUIUtility.labelWidth;
-
-				var typeRect = new Rect(position.x + 1, position.y + 1, width - 2, EditorGUIUtility.singleLineHeight);
-				var idRect = new Rect(typeRect.xMax + 4, typeRect.y, position.width - width - 4,
-				                      EditorGUIUtility.singleLineHeight);
-
-				var types = usam.latestManifest.availableFederationTypes.ToArray();
-				var selectedType = Array.IndexOf(types, federationTypeProp.stringValue);
-				var nextSelectedType = EditorGUI.Popup(typeRect, selectedType, types);
-				if (nextSelectedType < 0) nextSelectedType = 0;
-				federationTypeProp.stringValue = types[nextSelectedType];
-
-				var selectedId = options.IndexOf(federationIdProp.stringValue);
-				var nextSelectedId = EditorGUI.Popup(idRect, selectedId, options.ToArray());
-				if (nextSelectedId < 0) nextSelectedId = 0;
-				federationIdProp.stringValue = options[nextSelectedId];
-			}
+		private Dictionary<string, List<string>> GroupFederations(SerializedProperty setting)
+		{
+			var groups = new Dictionary<string, List<string>>();
 			
+			var entriesProperty = setting.FindPropertyRelative(nameof(BeamFederationSetting.entries));
+
+			for (int j = 0; j < entriesProperty.arraySize; j++)
+			{
+				var entry = entriesProperty.GetArrayElementAtIndex(j);
+				string federationId = entry.FindPropertyRelative(nameof(BeamFederationEntry.federationId)).stringValue;
+				string interfaceName = entry.FindPropertyRelative(nameof(BeamFederationEntry.interfaceName)).stringValue;
+
+				if (string.IsNullOrEmpty(federationId) || string.IsNullOrEmpty(interfaceName))
+					continue;
+
+				if (!groups.ContainsKey(federationId))
+				{
+					groups[federationId] = new List<string>();
+				}
+
+				groups[federationId].Add(interfaceName);
+			}
+
+			return groups;
 		}
 	}